Move router reducer to separate file
