+++ packages/next/src/shared/lib/app-router-context.ts
@@ -1,7 +1,7 @@
 'use client'
 
 import React from 'react'
-import type { FocusAndScrollRef } from '../../client/components/reducer'
+import type { FocusAndScrollRef } from '../../client/components/router-reducer/router-reducer'
 import type { FlightRouterState, FlightData } from '../../server/app-render'
 
 export type ChildSegmentMap = Map<string, CacheNode>
